Changed Makefile behavior so that the small Pecos test executables are NO LONGER LINKED as a result of a simple invocation of make.
Both Karen and Laura reported problems with the link step (which I've not
been able to reproduce).  Since DAKOTA testing is key tonight, I see no
reason to risk the build/testing not completing because potential issues
in Pecos test linkage.  The simple 'gmake' step will proceed since there's
nothing to do in the Pecos directory with this change.

Those of us testing Pecos will need to execute 'gmake check' to link the
test executables for the next day (or so).
I will restore the desirable behavior to this Makefile once we're confident
full builds/tests are completed without hitting this trivial "snag".



git-svn-id: svn+ssh://software.sandia.gov/svn/pecos/trunk@105 ccfab4c5-ad45-0410-a689-f63b6aebdeca