Switch to Mersenne Twister as default uniform random number generator (in contexts where the LHS uniform random number generator was being used).  This will cause some diffs in the nightly tests; for example, the first test (dakota_4level.in) shows a log of diffs that someone more familiar with this test than I should assess.  As a (temporary?) debugging measure, invoking Dakota with $DAKOTA_LHS_UNIFGEN="rnumlhs1" will cause the LHS uniform generator to be used (as it has been), e.g., as in the Bourne-shell invocation
	DAKOTA_LHS_UNIFGEN=rnumlhs1 dakota dakota_4level.in

I started to add keyword uniform_generator to dakota.input.nspec so
one could specify the choice of uniform random number generators in
.in files, but modifying constructors to specify this detail along
with the seed required changes to many files, so for now I've
abandoned this effort (while retaining copies of the files I had
modified).  I'm not convinced that adding this complexity would be
worthwhile, but of course we can always revisit this issue.  


git-svn-id: svn+ssh://software.sandia.gov/svn/pecos/trunk@178 ccfab4c5-ad45-0410-a689-f63b6aebdeca