Added configure option to specify --with-boost=<DIR>.
The default, '--with-boost=yes' will use the DAKOTA/Pecos customized Boost 1.37
version that is available as an SVN external in the packages subdirectory.
I do not recommend that anyone tries to use this new option to leverage other
Boost versions (for now -- it is strictly in the experimental phase).

Also, some minor cleanup in the --enable-debug settings.  I've removed the
--with-opt=<LEVEL> setting since my hacking of the CXXFLAGS in the M4 seemed
to make things worse.



git-svn-id: svn+ssh://software.sandia.gov/svn/pecos/trunk@212 ccfab4c5-ad45-0410-a689-f63b6aebdeca