Added a CMake build harness for Pecos.  Not quite functional, but I do not want to risk losing my local tweaks so committing a nearly-functional version (at least the libraries build on GNU Linux).
The CMake macros in the cmake directory are just a minimal set of the
"extra goodies" that have been added to support the Trilinos project.  Not sure
DAKOTA will want all of those used by Trilinos, so I'll probably gradually
add more as I see fit.

One notable surprise is there is currently no need for a "pecos_config.h".
This version obtains configuration information from Teuchos_config.h.  Probably
not good, but at least the silly "PACKAGE_NAME" macros are not redefined like
the situation that occurs with autotools.



git-svn-id: svn+ssh://software.sandia.gov/svn/pecos/trunk@272 ccfab4c5-ad45-0410-a689-f63b6aebdeca