Removed the "C wrapper" interface/functions coupling with LHS out of Pecos since they're now exposed by and defined in liblhs.a (or .dll/.so if you prefer).
Here's some justification that I received in an email from Kitware which is
probably worth logging:

>>>>> Brad King wrote:
My goal is to build Pecos in Visual Studio to demonstrate use of an
outside LHS built by GNU Fortran.  If everything links together
properly then the pecos_lhs_driver should pass.

I've hacked up the Pecos source from the above tarball to not enable
Fortran at all and instead import LHS from outside the build.  I got
all the way to linking the test executable but it failed with

  ../src/libpecos_src.so: undefined reference to 'lhs_dist2_'

because the "src/lhs_wrappers.f90" source is never compiled (since
Fortran is not enabled).  Clearly if we want Pecos to build in the
Visual Studio compiler without Fortran support then this wrapper
source cannot be built.

Is there a reason that lhs_wrappers.f90 is not built in LHS as part
of its API?  It seems like such convenient functions would be useful
to everyone using LHS.
<<<<<

OK'd by Mike Eldred and Laura Swiler.



git-svn-id: svn+ssh://software.sandia.gov/svn/pecos/trunk@404 ccfab4c5-ad45-0410-a689-f63b6aebdeca