CMake: Link pecos tests to lhs before lapack
LHS defines and uses at least one symbol (SSPEV) that is also provided
by lapack.  Put lhs on the link line first to ensure that its copy of
the symbol is preferred over the conflicting copy in lapack.  Otherwise
LHS may call the wrong copy of the symbol and crash.

git-svn-id: svn+ssh://software.sandia.gov/svn/pecos/trunk@466 ccfab4c5-ad45-0410-a689-f63b6aebdeca