Added files to implement compressed sensing algorithms. In addition to the Basis Pursuit and Orthogonal Matching Pursuit which were prevoulsy implemented, LASSO, Least angle regression, elastic net and Basis Pursuit denoising are now available.
The hpp and cpp files pf MathTools and LinearAlgebra are necessary for compressed sensing to compile and run. Test 66 in dakota_uq_rosenbrock_pce.in (added in a previous commit ) tests the construction of a PCE from an under-determined vandermonde which will call the compressed sensing solver. At the moment all previous logic applies. That is a switch uses L2 regression if the vandermonde is over-determined and compressed sensing if under-determined. The default compressed sensing solver is set to Lasso.

The next step is to provide logic for choosing the compressed sensing solver and its parameters through the input deck.


git-svn-id: svn+ssh://software.sandia.gov/svn/pecos/trunk@738 ccfab4c5-ad45-0410-a689-f63b6aebdeca