Checkpointing progress in activating the INTERPOLATION_OF_PRODUCTS option for nodal interpolants in partial integration scenarios ("all variables" mode). I want to make this option the default for v5.3 for three reasons: (1) consistency with all other Nodal and Hierarchical formulations, (2) access to gradient-enhancement modes, and (3) deactivation of the PRODUCT_OF_INTERPOLANTS_FAST option since it's     results, while seemingly accurate, remain suspect.
The issue with the previous implementation of INTERPOLATION_OF_PRODUCTS was
inaccuracy in the variance and variance gradient due to interpolation of the 
higher-order covariance function on the original grid (integration of covariance
on the original grid is fine, but interpolation is not).  The revised approach 
in this commit retains integration over the random dimensions on the original
grid, but interpolates the covariance function for the non-random dimensions
on a higher-order grid with ~twice the collocation points (while respecting
any point nestedness constraints as well as the level index-based basis 
management).  This approach is generating much more accurate results for both 
tensor and sparse grids, but I am not yet activating it as the default, pending
completion of some efficiency enhancements that are needed to hopefully prevent
nightly testing timeouts.

Supports Trac ticket #4218.


git-svn-id: svn+ssh://software.sandia.gov/svn/pecos/trunk@786 ccfab4c5-ad45-0410-a689-f63b6aebdeca