Tuning of barycentric interpolant evaluation.  Uses logic to split interpolation into 3 cases: (1) no interpolation (evaluation point matches a collocation point in every dimension) allowing simple lookup by collocation index, (2) full interpolation over all variables (no exact matches), and (3) partial interpolation (more than 1 exact match) where there is the option to collapse the interpolation into active points (3a: less looping but more bookkeeping) or leave it over all collocation points (3b: more looping but less bookkeeping).
For performance timing, I selected a level 5 sparse grid over 5 variables,
with evaluations at each collocation point when forming numerical expansion
moments and evaluations at 10^6 LHS samples for evaluating a probability 
mapping.  The most relevant timings are:
  v5.3:                 ~2100   CPU sec (no barycentric logic)
  current trunk version: ~770   CPU sec (second barycentric iteration)
  tuned w/ partial 3a:    599.8 CPU sec (average over several experiments)
  tuned w/ partial 3b:    597.3 CPU sec (average over several experiments)

Since the additional bookkeeping in 3a is not helping for this benchmark, 
this checkin has the simpler option 3b as the active option.  This is somewhat
expected given current use cases: partial interpolation should only occur for 
numerical expansion moment evaluation on sparse grids (where a tensor 
interpolant is evaluated for a point from another grid), and this occurrence 
is generally of secondary concern.


git-svn-id: svn+ssh://software.sandia.gov/svn/pecos/trunk@809 ccfab4c5-ad45-0410-a689-f63b6aebdeca