Correct Teuchos solver-related bug in trans_Z_to_U (and jacobian)
The final RealVector or RealMatrix these functions populate may be a Teuchos::View for updating in-place, for example a Dakota continuous_variables_view.  However, SerialDenseSolver::solve(), line 647 will call (LHS_ = u_vars) = RHS_, disconnecting the view.  

To work around, we use a temporary.  Note that the existing resize could also be problematic if the passed output object is a View.



git-svn-id: svn+ssh://software.sandia.gov/svn/pecos/trunk@1070 ccfab4c5-ad45-0410-a689-f63b6aebdeca