Change LHS interface to use Fortran 2003 ISO C bindings for function prototypes and string passing.
This was driven by newer compilers (in debug mode) checking the "hidden array length" parameter on the Fortran side, which was not populated by our C to Fortran interface.  

Working on a minimum change principle, I only modified C_wrap.f90 on the LHS side and LHSDriver.cpp on the Pecos side.  I left the caching and passing of fixed length strings in the C++ code, although it's not strictly needed any longer.  This leave the core of LHS unmodified.

We could go further to pass the C string as variable length and process the null-terminated C string in the Fortran.

If this approach proves portable, should do same for array passing from C to Fortran, and likely all C to Fortran interfaces.



git-svn-id: svn+ssh://software.sandia.gov/svn/pecos/trunk@1138 ccfab4c5-ad45-0410-a689-f63b6aebdeca