Orthogonal polynomial approximation based on a hierarchical surplus approach
Adds a pre-processing step within compute_coefficients() to offset the
SurrogateData according to the approximation values at lower levels.
Includes significant rework of the SurrogateData class to allow deep
copying of the SurrogateDataRep configured with either shallow or deep
copies of SurrogateData{Vars,Resp} instances.

TO DO: add communication of approximation level to compute_coefficients(),
as for {store,restore}_coefficients().