Promote combine type and discrepancy type to ExpansionConfigOptions
Promoting combine type to ExpansionConfigOptions eliminates the need to
pass the correction type through the combine() APIs and the need to cache
its value.  Promoting the discrepancy type to ExpansionConfigOptions relaxes
the need to detect recursion based on an index != _NPOS.

Initial pass at multiplicative discrepancy recursion, to be refined tomorrow.