Checkpointing work on refactoring stored levels and popped refinements
Consolidate and generalize storage to use a multi-index key for both
multilevel-multifidelity expansions and candidate refinements.  This
eliminates the store/restore workflow as now all expansions are available
for query, as keyed by model form/level.  This also allows candidate
refinements to co-exist across multiple levels, so that a greedy multilevel
scheme can compete level refinement sets against each other.

Creating a multilevel_harden branch in Pecos for this work in progress.
SurrogateData modifications are complete, and Shared*ApproxData changes
are in process.  *PolyApproximation and *Driver to come after that.