Surrogates: CMake generates Python wrappers in build dir
Re-enable the DAKOTA_PYTHON_API option and connect the distutils
setup.py build of Python wrappers of surrogates to the CMake build
system.

Python components will now get generated to the build tree into
build/python. When in build/python:
 * make pydakota --> setup.py build
 * make install  --> setup.py install
 * make clean    --> rm build swigpyrun.h

This removes the old CMake facilities for running SWIG to generate the
wrappers.