Separate out an anistopic version of update_smolyak_arrays()
A number of implementations had been fleshed out in IncrementalSparseGridDriver::
update_smolyak_arrays() which trade efficiency with the level of assumptions used
in the updating logic.  Previous logic was sufficient for uniform refinement but not
for anisotropic refinement.  Now that anisotopic refinement cases employ
update/increment  instead of re-allocation workflows, it was necessary to detect
anisotropic refinement and branch into the alternate logic.

Restores a number of tests.