Restucturing of push / finalize of drivers/ data / coefficients.
Note: this is moving a bit in the direction of destabilization of ML harden.
The stabilized revision for Dakota was fd0a8ae72daf079f68b15c6923f803b9f378f991.
Will need to evaluate whether to merge this or a subsequent revision at some
point in the future when the repo is reopened.