Consolidating {push,restore,finalize}_index() management within adaptive sparse grid drivers.
Separate push_index(), which is for internal use and may be flattened or
hierarchical as convenient, from restore_index() and finalize_index(), which
provide a consistent flattened definition that is published for external
clients (e.g., Dakota::SharedPecosApproxData).  All three sets of indices
are now cached in the Driver classes at push / finalize time in order to
simplify the pre_push/push/post_push and pre_finalize/finalize/post_finalize
logic in the SharedApproxData classes.  Work in progress.