Python surrogate wrap: resolve runtime type errors
Add a few explicit namespace qualifications so generated wrappers can
work with the passed datatypes.

This restores build to previous state of 5 known failures due to
non-implemented tests and 1--2 due to real runtime errors depending on
Python version.

TODO: Need to better understand the scope of namespace declarations
and using in SWIG wrappers.