Replace parametricUpdate flag with more persistent logic.
The state of the parametricUpdate flag was vulnerable to redundant updates,
which could cause failure in updating the Gauss points/weights after a
distribution parameter change.  New logic is based on whether colloc{Pts,Wts}
have been cleared, which is more robust to update_*() and update_from_*()
Model recursion updating.