DP-280: Use CMake imported targets for Boost
Pecos only uses Boost headers. Update the specification of link
libraries to use the CMake imported target Boost::boost (the Boost
headers).

Note that removal of Boost_INCLUDE_DIRS and related variables means
that it's crucial to use the Boost::boost target to make sure the
correct Boost headers are used when compiling, instead of a system
default Boost.