Bug fix branch: enforce recovery of first PCE coeff in OMP and LARS
Compressed sensing may fail to recover the mean coeff if it is small.
OMP already had a faciity to force recovery of certain indices, so that
feature is now active in Pecos::RegressOrthogPolyApproximation and has
been replicated for LARS.  Unfortunately, LASSO cannot easily support
this because it dynamically adds and removes coeffs.