cs_enforce_coeff0: simplify PCE regression logic by replacing sample mean,stdev conditionals with response scaling
Response scaling, implemented previously for regression of small discrepancy
response values, is simpler and should be sufficient for reliable recovery of
the mean coefficient.  This deactivates much of the previous commit, replacing
auto-magical data shifting with the optional user spec for response_scaling.
Should also bypass a set of baseline updates which would have been required
for the previous shifting approach.