Some hardening to allow for missing coeff 0 during sparse recovery
Since we will rely on the optional response_scaling to address loss of coeff 0
across the CS solvers, this coeff could still be lost when scaling is not
used.  This commit overrides additional OrthogPolyApproximation virtual fns
to provide general sparse implementations for mean(), combined_mean(), and
mean_gradient() for active and all views.